moved AppRouter and AuthContext into utils directory
