+++ neuromance/src/app/page.tsx
@@ -4,7 +4,7 @@ import dynamic from "next/dynamic";
 import styles from "./page.module.css";
 
 // Dynamically import your component to disable SSR
-const AppRouter = dynamic(() => import("./components/AppRouter/AppRouter"), {
+const AppRouter = dynamic(() => import("./utils/AppRouter/AppRouter"), {
   ssr: false,
 });
